requirements.txt
flask>=3.0
google-cloud-aiplatform>=1.60
